When handling SymbolicNames, the new SymbolicName constructor (not-committed yet in statcvs) must be used. There is a significant difference in the way CVS operates vs SVN. SVN tags by creating a revision which is a copy of another revision. Each checkin changes the 'revision' across the board, the log therefore may involve only a given number of files for this.
I have found a way to go through all files and associate the latest 
version of a given file with the appropriate SymbolicName.

However, the way the SymbolicNames works is that it picks up the date 
from a bunch of revision Data.
Since the copy is not associated with ANY file, it would lose that 
date unless it could be given to the SymbolicName,

Therefore I am suggesting a transparent change to SymbolicName to add 
a constructor that would accept such a date and on call to getDate 
would behaves as usual UNLESS there was a given non-null date.